Add Mojang Profile API client library
The library includes an implementation based on the JAX-RS Client API. A sample consumer of the library is also supplied in the form of a CLI application that can resolve UUIDs to profiles, and profile names to profiles/UUIDs.

The dependency fix-up post-processing previously implemented in the query library's build script is now extracted into a PomUtils utility class in the build script source directory to allow sharing of the implementation across sub-projects.