Implement Spring RestTemplate profile client
To allow for optional dependencies as with Maven, the Spring Framework's propdeps Gradle plugin is added to this project. This allows profile-spring to optionally depend on Jackson 2 to implement the default message converter for the SessionProfile class. To support automatic Jackson deserialisation of the Profile class, some field names have been updated to match the incoming JSON, resulting in corresponding updates to the JAX-RS implementation.

The profile-cli test application is also updated to support specifying, as an argument, the ProfileClient implementation to use.